+++ source/Public/Get-SqlDscRSConfiguration.ps1
@@ -96,7 +96,7 @@ function Get-SqlDscRSConfiguration
         $Version,
 
         [Parameter()]
-        [ValidateRange(1, [System.Int32]::MaxValue)]
+        [ValidateRange(1, [System.Int32]::MaxValue - 1)]
         [System.Int32]
         $RetryCount = 1,